+++ src/components/NodeProvider.jsx
@@ -1,88 +1,8 @@
 import PropTypes from "prop-types";
-import React, { createContext, useContext, useEffect, useState } from "react";
+import React from "react";
 
 import BoxedCircularProgress from "./BoxedCircularProgress";
-import { useLogger } from "./LoggerProvider";
-import { useSession } from "./SessionProvider";
-
-const NodeContext = createContext(null);
-
-function useStateOnce(callback) {
-  const [result, setResult] = useState(null);
-  const [processing, setProcessing] = useState(false);
-
-  useEffect(() => {
-    if (result === null && !processing) {
-      const callbackResult = callback();
-      if (callbackResult instanceof Promise) {
-        setProcessing(true);
-        callbackResult
-          .then((newResult) => {
-            setResult(newResult);
-          })
-          .finally(() => {
-            setProcessing(false);
-          });
-      } else {
-        setResult(callbackResult);
-      }
-    }
-  });
-
-  return result;
-}
-
-function useNode() {
-  return useContext(NodeContext);
-}
-
-function usePublisher(messageType, topicName) {
-  const node = useNode();
-  const logger = useLogger();
-
-  return useStateOnce(() => {
-    return node.createPublisher(messageType, topicName).catch((err) => {
-      logger.error(`Failed to create a new Publisher! ${err.message}`);
-    });
-  });
-}
-
-function useSubscription(messageType, topicName, callback) {
-  const node = useNode();
-  const logger = useLogger();
-
-  return useStateOnce(() => {
-    return node
-      .createSubscription(messageType, topicName, callback)
-      .catch((err) => {
-        logger.error(`Failed to create a new Subscription! ${err.message}`);
-      });
-  });
-}
-
-function useClient(serviceType, serviceName) {
-  const node = useNode();
-  const logger = useLogger();
-
-  return useStateOnce(() => {
-    return node.createClient(serviceType, serviceName).catch((err) => {
-      logger.error(`Failed to create a new Client! ${err.message}`);
-    });
-  });
-}
-
-function useService(serviceType, serviceName, callback) {
-  const node = useNode();
-  const logger = useLogger();
-
-  return useStateOnce(() => {
-    return node
-      .createService(serviceType, serviceName, callback)
-      .catch((err) => {
-        logger.error(`Failed to create a new Service! ${err.message}`);
-      });
-  });
-}
+import { NodeContext, useLogger, useSession, useStateOnce } from "../hooks";
 
 function NodeProvider({ children, nodeName }) {
   const logger = useLogger();
@@ -113,11 +33,4 @@ NodeProvider.propTypes = {
   nodeName: PropTypes.string.isRequired,
 };
 
-export {
-  NodeProvider,
-  useClient,
-  useNode,
-  usePublisher,
-  useService,
-  useSubscription,
-};
+export default NodeProvider;
