+++ src/components/examples/index.js
@@ -1,0 +1,4 @@
+export { default as SimpleClientNode } from "./SimpleClientNode";
+export { default as SimplePublisherNode } from "./SimplePublisherNode";
+export { default as SimpleServiceNode } from "./SimpleServiceNode";
+export { default as SimpleSubscriptionNode } from "./SimpleSubscriptionNode";
