+++ src/hooks/useNode.js
@@ -1,0 +1,9 @@
+import { createContext, useContext } from "react";
+
+const NodeContext = createContext(null);
+
+const useNode = () => {
+  return useContext(NodeContext);
+};
+
+export { NodeContext, useNode };
