+++ src/hooks/useSession.js
@@ -1,0 +1,9 @@
+import { createContext, useContext } from "react";
+
+const SessionContext = createContext(null);
+
+const useSession = () => {
+  return useContext(SessionContext);
+};
+
+export { SessionContext, useSession };
